♻️ 댓글을 수정시에 로그인 유저와 댓글을 쓴 유저가 같은지 검증로직 추가 view부분에서도 댓글 수정시에 수정됨을 알리는 동시에 수정된 시간을 뷰에 보일 수 있게 수정

+++ src/main/java/codesquad/codestagram/controller/CommentController.java
@@ -1,11 +1,11 @@
 package codesquad.codestagram.controller;
 
+import codesquad.codestagram.annotation.Login;
 import codesquad.codestagram.domain.Article;
 import codesquad.codestagram.domain.Comment;
 import codesquad.codestagram.domain.User;
 import codesquad.codestagram.service.ArticleService;
 import codesquad.codestagram.service.CommentService;
-import codesquad.codestagram.service.UserService;
 import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
@@ -26,12 +26,11 @@ public class CommentController {
         this.commentService = commentService;
         this.articleService = articleService;
     }
-
+    //댓글 등록
     @PostMapping("/articles/{articleId}")
-    public String addComment(@PathVariable("articleId") Long articleId,
+    public String addComment(@Login User loginUser, @PathVariable("articleId") Long articleId,
                            @RequestParam("content") String content,
                            HttpSession session) {
-        User loginUser = (User) session.getAttribute("loginUser");
         Article article = articleService.findArticleById(articleId)
                 .orElseThrow();
         
@@ -42,16 +41,19 @@ public class CommentController {
     }
 
     @PostMapping("/{articleId}/{commentId}/edit")
-    public String editComment(@PathVariable("articleId") Long articleId,
-                            @PathVariable("commentId") Long commentId,
-                            @RequestParam("content") String content) {
+    public String editComment(@Login User user, @PathVariable("articleId") Long articleId, @PathVariable("commentId") Long commentId, @RequestParam("content") String content) {
+        if(!user.equals(commentService.findCommentByCommentId(commentId).getUser())){
+            throw new IllegalStateException("댓글 수정 권한이 없습니다.");
+        }
         commentService.updateComment(commentId, content);
         return "redirect:/qna/articles/" + articleId;
     }
 
     @PostMapping("/{articleId}/{commentId}/delete")
-    public String deleteComment(@PathVariable("articleId") Long articleId,
-                                @PathVariable("commentId") Long commentId) {
+    public String deleteComment(@Login User user , @PathVariable("articleId") Long articleId, @PathVariable("commentId") Long commentId) {
+        if(!user.equals(commentService.findCommentByCommentId(commentId).getUser())){
+            throw new IllegalStateException("댓글 삭제 권한이 없습니다.");
+        }
         commentService.deleteComment(commentId);
         return "redirect:/qna/articles/" + articleId;
     }
